Backend/requirements.txt
Flask
Flask-CORS
requests
gunicorn
pandas
python-dotenv
gdown
numpy
scikit-learn